requirements.txt
python-dotenv==1.0.0
python-dateutil==2.8.2
google-api-python-client==2.94.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
